Update README.md
This project demonstrates how to automate report generation using Python by reading sales data from a CSV file, performing basic data analysis, creating visualizations, and exporting the results to a well-formatted PDF file. The script uses the pandas library to process the data and calculate total sales by department. A bar chart is then created using matplotlib to visually represent the sales distribution. Finally, the fpdf library is used to generate a professional-looking PDF report that includes the chart and a table summarizing the results. This automation reduces manual effort in generating recurring reports and can be customized for use in business, education, or research scenarios. The project includes the Python script, sample data, and a sample output report, and is structured to be easily extendable and adaptable for different types of data or report formats.